Extract OpenLegsCard badge helpers into a shared module
The covered-call screen at /positions/[id]/covered-call (issue #247) needs the
same coverage/dteBadge/pnl helpers OpenLegsCard already uses for the per-leg
breakdown table. Lifting them out keeps one source of truth for the visual
contract instead of duplicating it across two callsites.

Mechanical refactor — function bodies and the PILL_SHAPE constant move
byte-identical from OpenLegsCard into components/dashboard/openLegsBadges.js;
OpenLegsCard now imports them. Module lives in dashboard/ (not positions/) so
the import direction stays forward — positions → dashboard, never the reverse.

All 20 existing Playwright cases in dashboard-open-legs-card.spec.js pass
unchanged.
